more refactor cp
